backend/requirements.txt
# Core Django & REST Framework
Django>=5.0,<5.2
djangorestframework>=3.14,<3.16
djangorestframework-simplejwt>=5.3,<5.5
django-cors-headers>=4.3,<4.5
django-filter>=24.1,<25.0

# Database & Caching
psycopg2-binary>=2.9,<2.10
redis>=5.0,<5.3
django-redis>=5.4,<5.5
hiredis>=2.3,<3.0

# File & Media Handling
Pillow>=10.2,<11.0
# Cloudinary (Media Storage)
cloudinary>=1.36.0,<1.40.0
django-cloudinary-storage>=0.3.0,<0.4.0

# Authentication & Security
PyJWT>=2.8,<2.10
cryptography>=42.0,<43.0
argon2-cffi>=23.1,<24.0
django-ratelimit>=4.1,<4.2

# Background Tasks
celery>=5.3,<5.5
django-celery-beat>=2.6,<2.8
django-celery-results>=2.5,<2.6
flower>=2.0,<2.1

# Video Processing
opencv-python-headless>=4.9,<4.11
moviepy>=1.0,<1.1
python-ffmpeg>=2.0,<2.1

# Search & Analytics
elasticsearch>=8.12,<8.15
django-elasticsearch-dsl>=8.0,<8.1
elasticsearch-dsl>=8.12,<8.15

# SMS & Notifications
twilio>=9.0.0

sendgrid>=6.11,<6.12
firebase-admin>=6.4,<6.6
fcm-django>=2.0,<2.1

# Payments
stripe>=8.2,<9.0

# Monitoring & Logging
django-debug-toolbar>=4.3,<4.5
sentry-sdk>=1.40,<1.45
django-silk>=5.0,<5.2

# Data Validation & Serialization
marshmallow>=3.20,<3.22
python-dotenv>=1.0,<1.1
pydantic>=2.6,<2.8

# API Documentation
drf-spectacular>=0.27,<0.28
drf-yasg>=1.21,<1.22

# Date & Time
python-dateutil>=2.8,<2.9
pytz>=2024.1

# HTTP & Requests
requests>=2.31,<2.33
httpx>=0.26,<0.28

# Utilities
python-magic>=0.4,<0.5
python-slugify>=8.0,<8.1
shortuuid>=1.0,<1.1

# Testing & Quality
pytest>=8.0,<8.2
pytest-django>=4.8,<4.10
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<3.4
faker>=23.1,<24.0

# Performance
django-cachalot>=2.6,<2.7
django-compression-middleware>=0.5,<0.6
gunicorn>=21.2,<22.0
uvicorn[standard]>=0.27,<0.28
whitenoise>=6.6,<6.8

# Development
black>=24.1,<24.3
flake8>=7.0,<7.1
isort>=5.13,<5.14
pre-commit>=3.6,<3.7

# Environment
environs>=11.0,<12.0

# AI
Markdown>=3.5,<3.7

